backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
httpx==0.25.2
langchain==0.1.20
langchain-openai==0.1.0
langchain-community==0.0.38
langgraph==0.0.32
pydantic==2.5.0
aiofiles==23.2.1
yt-dlp==2023.12.30
youtube-transcript-api==0.6.2
pydub==0.25.1
